Splits AttackMove into trait and activity. Moves them to subfolders/namespaces accordingly.
