+++ OpenRA.Mods.RA/Activities/Hunt.cs
@@ -35,7 +35,7 @@ namespace OpenRA.Mods.RA.Activities
 				return this;
 
 			return Util.SequenceActivities(
-				new AttackMove.AttackMoveActivity(self, new Move.Move(self, target.Location, WRange.FromCells(2))),
+				new AttackMoveActivity(self, new Move.Move(self, target.Location, WRange.FromCells(2))),
 				new Wait(25),
 				this);
 		}
