+++ OpenRA.Mods.RA/Guard.cs
@@ -12,6 +12,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using OpenRA.Graphics;
+using OpenRA.Mods.RA.Activities;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.RA
@@ -36,7 +37,7 @@ namespace OpenRA.Mods.RA
 			self.SetTargetLine(target, Color.Yellow);
 
 			var range = WRange.FromCells(target.Actor.Info.Traits.Get<GuardableInfo>().Range);
-			self.QueueActivity(false, new AttackMove.AttackMoveActivity(self, self.Trait<IMove>().MoveFollow(self, target, WRange.Zero, range)));
+			self.QueueActivity(false, new AttackMoveActivity(self, self.Trait<IMove>().MoveFollow(self, target, WRange.Zero, range)));
 		}
 
 		public string VoicePhraseForOrder(Actor self, Order order)
